YO DAWG I HERD YOU LIKE HEADERS
* SO I'M not going to finish this dead meme because I already have a
theme going here.
* Updated the headers the "complete" version
* Also the edit that I tried to do in an amend in the first place
that once again started all this merge BS is in here as well, because
Git is dumb. Once again it only touched my branch, and is completely
irrelevant in terms of actual content, so IGNORE IT or I'm going to
slap you around a bit with a large trout.
* Also the beef needs some more attention.